+++ runtime/observability/telemetry-api/common/src/aws/smithy/kotlin/runtime/telemetry/metrics/Histogram.kt
@@ -10,7 +10,6 @@ import aws.smithy.kotlin.runtime.telemetry.context.Context
 import aws.smithy.kotlin.runtime.util.Attributes
 import aws.smithy.kotlin.runtime.util.emptyAttributes
 import kotlin.time.Duration
-import kotlin.time.ExperimentalTime
 import kotlin.time.measureTimedValue
 
 public interface Histogram<T : Number> {
@@ -51,7 +50,6 @@ public fun DoubleHistogram.recordSeconds(value: Duration, attributes: Attributes
  * @param context (Optional) trace context to associate with this measurement
  * @param block the code to execute and return a result from
  */
-@OptIn(ExperimentalTime::class)
 public inline fun <T> DoubleHistogram.measureSeconds(attributes: Attributes = emptyAttributes(), context: Context? = null, block: () -> T): T {
     val tv = measureTimedValue(block)
     recordSeconds(tv.duration, attributes, context)