+++ runtime/smithy-client/common/src/aws/smithy/kotlin/runtime/client/endpoints/Endpoint.kt
@@ -52,14 +52,7 @@ public data class Endpoint @InternalApi constructor(
         other is Endpoint &&
             uri == other.uri &&
             headers == other.headers &&
-            attributesEqual(other)
-
-    private fun attributesEqual(other: Endpoint): Boolean =
-        attributes.keys.size == other.attributes.keys.size &&
-            attributes.keys.all {
-                @Suppress("UNCHECKED_CAST")
-                attributes.contains(it) && attributes.getOrNull(it as AttributeKey<Any>) == other.attributes.getOrNull(it)
-            }
+            attributes == other.attributes
 
     override fun hashCode(): Int {
         var result = uri.hashCode()